Add availability page Added feature - employees current availability loads into the display when they are selected
Now checks if emp has availalbity then parses the data and sets the check boxes on the view to reflect this.

When a new employee is selected the flags are all reset to false and then set again to reflect the newly selected employee.

Main methods added to AvailabilityController  loadEmp, loadAvailability