LATEST: added comments for disabling certain time slots
Figured the reason why it only went up to Friday Add Availability disable checkboxes, it was because in the FXML the date labels were inbetween Friday checkboxes and Saturday checkboxes. This shouldn't have been an issue as in the Grid pane, the checkboxes were still given the correct corresponding grid numbers. Weird.

Example businessID of '4' used as unsure how to retrieve current business logged in.

PROBLEM: Keeps inserting the same data into the businessDetails database page and the userLogin database page.